// apps/web/src/types/generated/MonthlyReportSchedule.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MonthlyReportDelay } from "./MonthlyReportDelay";
import type { MonthlyReportPreset } from "./MonthlyReportPreset";

export type MonthlyReportSchedule = {
  "Preset": { preset: MonthlyReportPreset; delay: MonthlyReportDelay };
} | {
  "Custom": { day: number; second_of_day: number; delay: MonthlyReportDelay };
};